media-source/SourceBuffer-abort.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Check the values of appendWindowStart and appendWindowEnd after abort()</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        var sourceBuffer = mediaSource.addSourceBuffer(mime);
        sourceBuffer.abort();
        t.step(function() {
            assert_equals(sourceBuffer.appendWindowStart, 0);
            assert_equals(sourceBuffer.appendWindowEnd, Number.POSITIVE_INFINITY);
        });
        t.done();
    });
    video.src = window.URL.createObjectURL(mediaSource);
}, 'SourceBuffer#abort() : Check the values of appendWindowStart and appendWindowEnd.');

</script>
</body>
</html>
